Decrease Google Analytics importance to Low
As discussed by others in #405, there's no good reason for Google Analytics being in a category where it:
> can't be omitted by any reason.

To add on to what others have said, this statement above is also misleading. There shouldn't be a requirement to fork over user data to Google as part of all, or even most, front end development.